Fix SaltStack repo installation
The SaltStack repository site was tidied up a few weeks ago and
installations from this repository will have been failing ever since,
as the el7 repositories were relocated. Fix up the URLs and make them
a little more prominent in the install script so they're easier to
modify in the future.